Allow directories to be used in config
Variety is fun and some users, such as myself, probably like to have several prerolls for their daily rotation or certain times.
Adding the is_directory_check function allows us to determine whether a given path in the prerolls.json config is a file or directory and adjust behavior accordingly.
If the path is a file we pass it along as normal, but if it is a directory we convert the files into a semicolon-delimited string and use that for new_preroll. Also incidentally allows use of any single file for DAILYPATH regardless of name.

Other than making it slightly harder to parse logging if new_preroll string is very long, existing behavior should not otherwise be impacted.